+++ mobile/android/android-components/components/browser/icons/src/main/java/mozilla/components/browser/icons/pipeline/IconResourceComparator.kt
@@ -42,6 +42,9 @@ internal object IconResourceComparator : Comparator<IconRequest.Resource> {
 @Suppress("MagicNumber")
 private fun IconRequest.Resource.Type.rank(): Int {
     return when (this) {
+        // An icon from our "tippy top" list should always be preferred
+        IconRequest.Resource.Type.TIPPY_TOP -> 20
+
         // We prefer touch icons because they tend to have a higher resolution than ordinary favicons.
         IconRequest.Resource.Type.APPLE_TOUCH_ICON -> 15
         IconRequest.Resource.Type.FAVICON -> 10